requirements.txt
httpx==0.27.0
pydantic==2.7.1
pandas==2.2.2
sqlalchemy==2.0.30
streamlit==1.55.00
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.0
pytest-vcr==1.0.2
pytz==2024.1